+++ srcs/ft_mat_calc.c
@@ -6,11 +6,11 @@
 /*   By: mmerabet <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/04 20:50:53 by mmerabet          #+#    #+#             */
-/*   Updated: 2018/02/07 21:20:25 by mmerabet         ###   ########.fr       */
+/*   Updated: 2018/02/08 18:03:40 by mmerabet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ft_math.h"
+#include "ft_matrix.h"
 
 t_mat	*ft_mat_calc(t_mat a, t_mat b, t_mat *res, t_calcmode op)
 {
@@ -23,7 +23,7 @@ t_mat	*ft_mat_calc(t_mat a, t_mat b, t_mat *res, t_calcmode op)
 	return (ft_mat_add(a, b, res));
 }
 
-t_mat	*ft_mat_calci(t_mat a, double b, t_mat *res, t_calcmode op)
+t_mat	*ft_mat_calci(t_mat a, float b, t_mat *res, t_calcmode op)
 {
 	if (op == OP_SUB)
 		return (ft_mat_subi(a, b, res));
@@ -34,7 +34,7 @@ t_mat	*ft_mat_calci(t_mat a, double b, t_mat *res, t_calcmode op)
 	return (ft_mat_addi(a, b, res));
 }
 
-t_mat	*ft_mat_icalc(double a, t_mat b, t_mat *res, t_calcmode op)
+t_mat	*ft_mat_icalc(float a, t_mat b, t_mat *res, t_calcmode op)
 {
 	if (op == OP_SUB)
 		return (ft_mat_isub(a, b, res));
